+++ src/main/java/com/anemone/bot/receiver/BotEventListener.java
@@ -20,7 +20,6 @@ import jakarta.annotation.PostConstruct;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
 
 /**
  * Shiro 事件监听器